Add transparent support for count query embargo on failure. (#294)
* Add transparent support for count query embargo on failure.

* changes as per review

Fixes #161
Fixes #164 
Fixes #226 
Fixes #236 
Fixes #254 
Fixes #286 
Fixes #287
